requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
urllib3==1.26.16
gigachat==0.1.8
python-jose==3.3.0 